Modificado Javi

+++ src/main/java/meta/utils/GeneradorLog.java
@@ -31,7 +31,7 @@ public class GeneradorLog {
 
         }
     }
-    public void escritura(String algoritmo,String funcion, String semilla,long tiempoInicial,long tiempoFinal, double[] solucion){ //aqui añadimos el texto que queramos mostrar
+    public void escritura(String algoritmo,String funcion, String semilla,double tiempoInicial,double tiempoFinal, double[] solucion){ //aqui añadimos el texto que queramos mostrar
         String cadena= "algoritmo";
         texto.append(cadena);
